Make sccache_version a simple string variable
Defining the SHA256 hashes in the role defaults made this role a bit of
a pain to maintain, since whenever a new sccache version was released,
we'd need to look up all these checksums by hand and make a new release
of the role. Since the sccache project also publishes checksum files
alongside the release binaries, it's a bit redundant since the `get_url`
Ansible module can simply fetch these files for us.